requirements.txt
numpy
opencv_python
pandas
Pillow
requests
streamlit
streamlit_lottie
streamlit_option_menu
tensorflow